Refactored entire file structure, added handy dandy MenuView widget which can quickly make a menu page, Also made an app bar widget, that avoids the notch problem as SafeArea was nopt working, to use it simply place the app bar at the top of a column containing your page contents
