+++ lib/pages/SortingAlgScreen.dart
@@ -1,0 +1,17 @@
+import 'package:algo_io_files/util/menu_view.dart';
+import 'package:flutter/material.dart';
+
+class SortingAlgScreen extends StatelessWidget {
+  final List<String> _topics = [
+    'Merge Sort',
+    'Quick Sort',
+  ];
+
+  final String _title = "Sorting Algorithms";
+
+  @override
+  Widget build(BuildContext context) {
+    return MenuView(_title, _topics);
+  }
+}
+
